Atualização dados do Banco (add password a user)

+++ src/database/migrations/1635949766368-CreateUsers.ts
@@ -17,7 +17,7 @@ export class CreateUsers1635949766368 implements MigrationInterface {
                 },
                 {
                     name: "matricula",
-                    type: "integer"
+                    type: "bigint"
                 },
                 {
                     name: "cargo",
